parallel-demo/pyproject.toml
[project]
name = "parallel_demo"
version = "0.1.0"
description = "CrewAI 并行执行演示项目"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.1.0",
    "crewai-tools>=0.1.0",
    "pysqlite3-binary>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/parallel_demo"]
